feat: handle row interaction | sonar fixes - commonise code
TISNEW-4834
